chore: Final folder structure and minor reorganization of files/classes/modules
