--- src/drawthis/services/settings/folder_service.py
@@ -1,1 +1,0 @@
-""" """
