+++ src/iteradraw/application/commands/timer_commands.py
@@ -1,0 +1,11 @@
+from dataclasses import dataclass
+
+
+@dataclass(frozen=True)
+class AddTimerCommand:
+    duration: int
+
+
+@dataclass(frozen=True)
+class RemoveTimerCommand:
+    duration: int
