+++ src/iteradraw/application/services/slideshow_service.py
