+++ src/iteradraw/domain/exceptions.py
