+++ src/iteradraw/infrastructure/filesystem/crawler.py
