+++ src/iteradraw/infrastructure/graphics/feh_backend.py
