+++ src/iteradraw/presentation/tkinter_legacy/tkinter_viewmodel.py
