+++ src/iteradraw/shared/types.py
